second commit

+++ ApiiCollabdev/src/main/java/com/example/ApiiCollabdev/Repository/ContributionRepository.java
@@ -3,5 +3,5 @@ package com.example.ApiiCollabdev.Repository;
 import com.example.ApiiCollabdev.entities.Contribution;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContributionRepository extends JpaRepository<Contribution, Long> {
+public interface ContributionRepository extends JpaRepository<Contribution, Integer> {
 }
